convert cell format view to vue 3 mount
